displaying 'game over' message when collided, deleting console logs
