+++ index.html
@@ -7,17 +7,13 @@
     <title>Snake Game!</title>
 </head>
 <body>
-<h1>Snake Game!</h1><span id="gameScore"></span>
 <span><button id="gameStart">Start the game!</button></span>
-
-<style>
-  .bg {
-  }
-
-</style>
+<link rel = "stylesheet"
+      type = "text/css"
+      href = "styles.css" />
 
 <div class="bg-wrapper">
-    <img src="images/bg.jpg" class="bg">
+
 </div>
 
 <canvas width="800px" height="800px" id="gameboard"></canvas>
